moving students and groups firebase functions to separate files
