+++ src/firebase/firebase.js
@@ -12,6 +12,8 @@ export const auth = firebase.auth();
 export { getAllRegistrations, writeRegistration } from './registrations';
 export { getLessons, getDateLessons, writeLessons } from './lessons';
 export { getAllAbsentees } from './absence';
+export { getGroups, createGroup } from './groups';
+export { deleteStudent } from './students';
 
 export async function getUsersRegistrations() {
   const querySnapshot = await db.collection("registrations").where("email", "==", getUserEmail()).get();
@@ -20,50 +22,6 @@ export async function getUsersRegistrations() {
     registrations.push(doc.data());
   });
   return registrations;
-}
-
-export async function createGroup(group) {
-  const userIsModerator = await checkUserClaim('moderator');
-  if (!userIsModerator) {
-    return new Error('User not authorized.');
-  }
-  return db.collection("groups").doc(group.groupName).set(group)
-    .then(()=> {
-      return {success: true};
-    })
-    .catch((error)=> {
-      throw new Error(error);
-    });
-}
-
-export async function deleteStudent(student) {
-  const studentDocName = getStudentDocName(student);
-  await db.collection('registrations').doc(studentDocName).delete();
-  const allGroups = await getGroups();
-  await removeStudentFromGroups(student, allGroups);
-}
-
-export async function getGroups() {
-  const querySnapshot = await db.collection("groups").get();
-  let groups = [];
-  querySnapshot.forEach((doc) => {
-    groups.push(doc.data());
-  });
-  return groups;
-}
-
-export async function removeStudentFromGroups(student, groups) {
-  const userIsModerator = await checkUserClaim('moderator');
-  if (!userIsModerator) {
-    return new Error('User not authorized.');
-  }
-  const studentDocName = getStudentDocName(student);
-  groups.forEach((group) => {
-    db.collection('groups').doc(group.groupName).collection('students').doc(studentDocName).delete()
-      .catch((error) => {
-        return new Error("Error removing document: ", error);
-      });
-  });
 }
 
 export async function writeStudentToGroup(student, groupName) {
